server_code/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
pymysql
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets